Fleshing out PXClient some more
